Update README.md
The step-by-step framework has two essential and distinguished phases:
1.)First step phase emphasizes the elaboration of the summarization feature and the capability to summarize financial-related text and content both from selected web pages and from financial news-related API-es using pre-trained language models like T5 and BERT.
2.)Second step phase enhances the sentiment analysis for each summarization and outlines the sentiment eather is positive, negative, or neutral.
The objective is to build a financial news summarization and sentiment analysis pipeline, step-by-step starting with the 1.)Firts step phase and then following the 2.)Second step phase. The result shall be to generate concise, accurate, and informative summaries of financial news articles.